+++ clox/params.py
@@ -58,3 +58,5 @@ FACES_MAP = {
 FACES_LIST = [-1] + sorted(FACES_MAP)
 
 CALENDAR_LIST = ["month", "year"]
+
+DATE_SYSTEMS_LIST = ["gregorian", "jalali"]