+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `--date-system` argument
+### Changed
+- `README.md` updated
 ## [0.6] - 2025-02-25
 ### Added
 - `--calendar` argument